Split Router into seperate classes
