+++ routes/init.rb
@@ -1,7 +1,7 @@
 require "sinatra/activerecord"
-require_relative "user"
-require_relative "login"
-require_relative "main"
-require_relative "signup"
-require_relative "images"
+require_relative "user_router"
+require_relative "session_router"
+require_relative "main_router"
+require_relative "signup_router"
+require_relative "image_router"
 require_relative "../app"
